fix bug with witness

+++ zkboo/prover.h
@@ -41,7 +41,7 @@ class Proof {
         CircuitComm comms[3];
         CircuitView *views[2];
         RandomSource rands[2];
-        uint8_t *w[2];
+        uint32_t *w[2];
         uint32_t *outShares[2];
         uint8_t *out;
         int output;
